Several bugfixes in the trade model.
