+++ core/input/generisdata_trade.prn
@@ -13,16 +13,16 @@ omv         Variable operation and maintenance costs given in $(2015)/kWa energy
 omv_d       Variable operation and maintenance costs given in $(2015)/kWa energy production. Per 1000km.s
 lifetime    Given in years
 $offtext
-+                        gas_pipe                 lng_liq                  lng_ves                 lng_gas
-tech_stat                     0                        0                        0                       0
-inco0                        20                      167                        0                       0.1
-inco0_d                     100                        0                        0                       0
-constrTme                     3                        3                        0                       3
-eta                        0.99                     0.99                        0.99                    0.99
-eta_d                      0.99                     0.99                        0.99                    0.99
-omf                        0.01                      3.3                        0                       0
-omf_d                         0                        0                        0                       0
-omv                           0                        0                        0                       0
-omv_d                         0                        0                        2.3                     0
-lifetime                     35                       35                        999                    35
++                        gas_pipe                 lng_liq                  lng_ves                 lng_gas              coal_ves
+tech_stat                     0                        0                        0                       0                     0
+inco0                        20                      167                        0                     0.1                     0
+inco0_d                     100                        0                        0                       0                     0
+constrTme                     3                        3                        0                       3                     0
+eta                        0.99                     0.99                     0.99                    0.99                  0.99
+eta_d                      0.99                     0.99                     0.99                    0.99                  0.99
+omf                        0.01                      3.3                        0                       0                     0
+omf_d                         0                        0                        0                       0                     0
+omv                           0                        0                        0                       0                     0
+omv_d                         0                        0                      2.3                       0                   2.3
+lifetime                     35                       35                      999                      35                   999
 *** EOF ./core/input/generisdata_trade.prn
