+++ modules/24_trade/capacity/datainput.gms
@@ -50,12 +50,14 @@ PARAMETERS
       / gas_pipe 0.0
         lng_liq 0.020
         lng_gas 100.0
-        lng_ves 100000.0 /
+        lng_ves 999999.0
+        coal_ves 999999.0 /
   p24_cap_relMaxGrowthRate(teTrade)                                             'Relative maximum yearly growth rate for trade transportation capacity (percent)'
       / gas_pipe 0.0
         lng_liq 0.01
         lng_gas 0.03
-        lng_ves 100000.0 /
+        lng_ves 999999.0
+        coal_ves 999999.0 /
 ;
 
 PARAMETER TABLE p24_disallowed(all_regi,all_regi,tradeModes)                    'Trade routes that are explicitly disallowed.'
